feat: add dedicated icon components and types
