+++ src/constants/external-icons.ts
@@ -1,0 +1,6 @@
+import { Icon } from "@/utils/types";
+
+export const JavaIcon: Icon = {
+	title: "Java",
+	dark: "/icons/icon-java-dark.svg",
+};
