--- src/constants/technologies.ts
@@ -1,16 +1,0 @@
-export enum Language {
-	Java = "Java",
-	Dart = "Dart",
-	JavaScript = "JavaScript",
-	TypeScript = "TypeScript",
-}
-
-export enum Tech {
-	React = "React",
-	Next = "Next.js",
-	TailwindCSS = "Tailwind CSS",
-	Firebase = "Firebase",
-	Flutter = "Flutter",
-	ShadcnUI = "shadcn/ui",
-	Supabase = "supabase",
-}
