+++ src/utils/helpers.ts
@@ -1,0 +1,6 @@
+import { Icon } from "@/utils/types";
+import { SimpleIcon } from "simple-icons";
+
+export const isSimpleIcon = (icon: SimpleIcon | Icon): icon is SimpleIcon => {
+	return (icon as SimpleIcon).svg !== undefined;
+};
