+++ src/utils/types.ts
@@ -1,0 +1,4 @@
+export interface Icon {
+	title: string;
+	dark: string;
+}
